requirements.txt
torch>=2.0
torchvision
diffusers
accelerate
controlnet-aux
transformers
monai
scikit-image
pillow
tifffile
numpy
pandas
openpyxl
scipy
pytorch-fid
matplotlib
seaborn
pyyaml
jupyter
